+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * Extract, Extract3 opcodes
     * ExtractUint16, ExtractUint32, ExtractUint64 opcodes
     * Txn, Global fields
+    * Added application account (a smart contract now has an escrow account). Updated checkpoint structure to store `applicationAccount` while running `algob` scripts.
 
 ### Breaking changes